refactor: 토큰 관련 로직을 추상클래스로 분리하고, 필요한 값을 넣어주는 식으로 해서 의존관계를 최소화
의존관계를 최소화 하며 토큰관련 로직은 로직을 제공하는 추상클래스로 구현
