+++ src/main/java/gift/service/auth/AuthService.java
@@ -1,13 +1,11 @@
 package gift.service.auth;
 
 import gift.model.Member;
-import gift.model.MemberRole;
+import gift.utils.AuthUtils;
 import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.security.Keys;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 
-import javax.crypto.SecretKey;
 import java.util.Date;
 
 @Service
@@ -28,32 +26,8 @@ public class AuthService {
                 .claim("role", member.getRole())
                 .issuedAt(new Date())
                 .expiration(new Date(System.currentTimeMillis() + expiredTime))
-                .signWith(getSecretKey())
+                .signWith(AuthUtils.getSecretKey(secretKey))
                 .compact();
         return token;
-    }
-
-    public Long getMemberIdWithToken(String token) {
-        var id = Jwts.parser()
-                .verifyWith(getSecretKey())
-                .build()
-                .parseSignedClaims(token)
-                .getPayload()
-                .getSubject();
-        return Long.parseLong(id);
-    }
-
-    public MemberRole getMemberRoleWithToken(String token) {
-        var role = Jwts.parser()
-                .verifyWith(getSecretKey())
-                .build()
-                .parseSignedClaims(token)
-                .getPayload()
-                .get("role");
-        return MemberRole.valueOf(role.toString());
-    }
-
-    private SecretKey getSecretKey() {
-        return Keys.hmacShaKeyFor(secretKey.getBytes());
     }
 }
